restore -std=c++98 flag to compil, get rid of stoul, creation of strToLong

+++ includes/Request.hpp
@@ -25,7 +25,7 @@ private:
 	int									_inSocket;
 	std::string							_buff;
 
-	size_t								_contentLength;
+	long								_contentLength;
 	std::string							_method;
 	std::string							_target;
 	std::string							_httpVersion;
